requirements.txt
openai
zhipuai
numpy~=1.26.4
python-dotenv~=0.21.0
torch
torchvision
torchaudio
transformers~=4.37.2
tqdm~=4.66.2
PyPDF2
markdown
html2text
tiktoken~=0.6.0
beautifulsoup4~=4.12.3
pydantic
gradio==4.0
pdfminer.six
pdf2image
unstructured
pillow_heif
opencv-python
pikepdf
pypdf
weaviate-client
langchain-openai
faiss-cpu
langchain~=0.1.14
requests~=2.31.0
streamlit
pandas~=2.2.1
python-docx
langchain-core>=0.1.0
langchain-community>=0.0.10
sentencepiece
